Move some files from components to hooks
