+++ optuna_dashboard/ts/components/StudyList.tsx
@@ -35,12 +35,12 @@ import { styled } from "@mui/system"
 import { StudySummary } from "ts/types/optuna"
 import { actionCreator } from "../action"
 import { useConstants } from "../constantsProvider"
+import { useCreateStudyDialog } from "../hooks/useCreateStudyDialog"
+import { useDeleteStudyDialog } from "../hooks/useDeleteStudyDialog"
+import { useRenameStudyDialog } from "../hooks/useRenameStudyDialog"
 import { studySummariesLoadingState, studySummariesState } from "../state"
 import { useQuery } from "../urlQuery"
 import { AppDrawer } from "./AppDrawer"
-import { useCreateStudyDialog } from "./CreateStudyDialog"
-import { useDeleteStudyDialog } from "./DeleteStudyDialog"
-import { useRenameStudyDialog } from "./RenameStudyDialog"
 
 export const StudyList: FC<{
   toggleColorMode: () => void
