+++ optuna_dashboard/ts/hooks/useDeleteStudyDialog.tsx
